connectingPost&Users&addingAuthChecks&ClearingPostUserRelations

+++ controllers/feed.js
@@ -4,6 +4,8 @@ const path = require('path');
 const { validationResult } = require('express-validator/check')
 
 const Post = require('../models/post')
+const User = require('../models/user')
+
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -34,47 +36,53 @@ exports.getPosts = (req, res, next) => {
       });
   };
 
-exports.createPost = (req,res,next) => {
-    const errors = validationResult(req)
-    if(!errors.isEmpty){
-        const error = new Error('Validation failed, entered data is incorrect');
-        error.statusCode = 422;
-        throw error;
+  exports.createPost = (req, res, next) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      const error = new Error('Validation failed, entered data is incorrect.');
+      error.statusCode = 422;
+      throw error;
     }
-    if(!req.file){
-        const error = new Error('No image provided.');
-        error.statusCode = 422;
-        throw error;
+    if (!req.file) {
+      const error = new Error('No image provided.');
+      error.statusCode = 422;
+      throw error;
     }
-    //const imageUrl = req.file.path;
     const imageUrl = req.file.path.replace("\\" ,"/");
+  
     const title = req.body.title;
     const content = req.body.content;
-    
+    let creator;
     const post = new Post({
-        title: title,
-        content: content,
-        imageUrl: imageUrl,
-        creator: {
-            name: 'Iza'
-        },
+      title: title,
+      content: content,
+      imageUrl: imageUrl,
+      creator: req.userId
     });
-
-    post.save()
-    .then(result=>{
-        console.log(result)
+    post
+      .save()
+      .then(result => {
+        return User.findById(req.userId);
+      })
+      .then(user => {
+        creator = user;
+        user.posts.push(post);
+        return user.save();
+      })
+      .then(result => {
         res.status(201).json({
-            message: 'Post created successfuly!',
-            post: result
+          message: 'Post created successfully!',
+          post: post,
+          creator: { _id: creator._id, name: creator.name }
         });
-    })
-    .catch(err=>{
-        if(!err.statusCode){
-            err.statusCode = 500;
+      })
+      .catch(err => {
+        if (!err.statusCode) {
+          err.statusCode = 500;
         }
         next(err);
-    });
-};
+      });
+  };
 
 exports.getPost = (req,res,next) => {
     const postId = req.params.postId;
@@ -126,6 +134,11 @@ exports.updatePost = (req, res, next) => {
           error.statusCode = 404;
           throw error;
         }
+        if(post.creator.toString()!==req.userId){
+          const error = new Error('Not authorized');
+          error.statusCode = 403;
+          throw error;
+        }
         if (imageUrl !== post.imageUrl) {
           clearImage(post.imageUrl);
         }
@@ -154,13 +167,25 @@ exports.updatePost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
           }
+          if(post.creator.toString()!==req.userId){
+            const error = new Error('Not authorized');
+            error.statusCode = 403;
+            throw error;
+          }
           //Chcek logged in user
           clearImage(post.imageUrl)
           return Post.findByIdAndRemove(postId);
       })
       .then(result=>{
-          console.log(result);
-          res.status(200).json({message: 'Deleted post.'})
+        return User.findById(req.userId)
+          
+      })
+      .then(user=>{
+        user.posts.pull(postId)
+        return user.save()
+      })
+      .then(result => {
+        res.status(200).json({message: 'Deleted post.'})
       })
       .catch(err=>{
           if(!err.statusCode){
